requirements.txt
asgiref==3.9.1
attrs==25.3.0
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.7.0
django_neomodel==0.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
neo4j==5.19.0
neomodel==5.3.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0